Externalize working tasks
