+++ deployments.json
@@ -1,52 +1,1 @@
-{
-  "ganache": {
-    "Storage": {
-      "TEST0": {
-        "name": "Storage",
-        "tag": "TEST0",
-        "address": "0x2612Af3A521c2df9EAF28422Ca335b04AdF3ac66",
-        "timestamp": "2023-12-09T09:52:32.000Z",
-        "transactionHash": "0x40225c27cad61874ac6fdf523b6989d608ab968af341e545eab0b142cd89b8d9",
-        "blockHash": "0xfddf417cf88db09ab9db60845c5dccce1336e229268b343ab1b25ea17dcb8ee6",
-        "codeHash": "0x83c80822e4dc1b4cbe735b05dd39547fbe4958cf1d12d48536bfbf793f27f82a",
-        "chainId": 1337,
-        "upgradeable": false
-      },
-      "TEST1": {
-        "name": "Storage",
-        "tag": "TEST1",
-        "address": "0x26b4AFb60d6C903165150C6F0AA14F8016bE4aec",
-        "timestamp": "2023-12-09T09:56:36.000Z",
-        "transactionHash": "0x75d429ea4d8ca0f4f057b191993976f85b4b14b87124c1116d170572d6c04e83",
-        "blockHash": "0xb28d288b79e4776bfe553e598300c0db8effb224bdb4a2d1c2ca4f731cdadb49",
-        "codeHash": "0x24548c07a3a751bf60f8e50b33c2d1da7bfb783f239885ada2d8af90e9ba3b2c",
-        "chainId": 1337,
-        "upgradeable": false
-      }
-    },
-    "StorageUpgr": {
-      "TEST0": {
-        "name": "StorageUpgr",
-        "tag": "TEST0",
-        "address": "0x0E696947A06550DEf604e82C26fd9E493e576337",
-        "logic": "0xDb56f2e9369E0D7bD191099125a3f6C370F8ed15",
-        "logicHistory": [
-          {
-            "timestamp": "2023-12-09T09:58:05.000Z",
-            "logic": "0x630589690929E9cdEFDeF0734717a9eF3Ec7Fcfe"
-          },
-          {
-            "logic": "0xDb56f2e9369E0D7bD191099125a3f6C370F8ed15",
-            "timestamp": "2023-12-09T09:58:27.000Z"
-          }
-        ],
-        "timestamp": "2023-12-09T09:58:05.000Z",
-        "transactionHash": "0xe8f836d10573ab4ac466305c820b527af7677d2fcdbe65be65d6db1e31cfad07",
-        "blockHash": "0xd666abd70256c53b2162aef4f6d7992e1aa9bce0b3ef0e7888c6d8ed3ab0a845",
-        "codeHash": "0x254c18f8ba0d25028d636b2d09643aa75362ffa5353169785c34df69a5974b46",
-        "chainId": 1337,
-        "upgradeable": true
-      }
-    }
-  }
-}
+{"ganache":{"Storage":{"TEST0":{"name":"Storage","tag":"TEST0","address":"0x5017A545b09ab9a30499DE7F431DF0855bCb7275","timestamp":"2023-12-09T12:07:03.000Z","transactionHash":"0xcd71f90e36a801b26720c19a7ffc6c600a6aa4d0c6b3272ecf7769cc521ddaa7","blockHash":"0xba5d56385a2b1dcc6f43ed8b5a74e75588df413d8ce2b2a9e608ea9cf0bb4323","codeHash":"0x24548c07a3a751bf60f8e50b33c2d1da7bfb783f239885ada2d8af90e9ba3b2c","chainId":1337,"upgradeable":false},"TEST1":{"name":"Storage","tag":"TEST1","address":"0x26b4AFb60d6C903165150C6F0AA14F8016bE4aec","timestamp":"2023-12-09T09:56:36.000Z","transactionHash":"0x75d429ea4d8ca0f4f057b191993976f85b4b14b87124c1116d170572d6c04e83","blockHash":"0xb28d288b79e4776bfe553e598300c0db8effb224bdb4a2d1c2ca4f731cdadb49","codeHash":"0x24548c07a3a751bf60f8e50b33c2d1da7bfb783f239885ada2d8af90e9ba3b2c","chainId":1337,"upgradeable":false}},"StorageUpgr":{"TEST0":{"name":"StorageUpgr","tag":"TEST0","address":"0xFF6049B87215476aBf744eaA3a476cBAd46fB1cA","logic":"0x86072CbFF48dA3C1F01824a6761A03F105BCC697","logicHistory":[{"timestamp":"2023-12-09T12:07:53.000Z","logic":"0x86072CbFF48dA3C1F01824a6761A03F105BCC697"},{"logic":"0xA586074FA4Fe3E546A132a16238abe37951D41fE","timestamp":"2023-12-09T12:08:24.000Z"},{"logic":"0x86072CbFF48dA3C1F01824a6761A03F105BCC697","timestamp":"2023-12-09T12:09:31.000Z"}],"timestamp":"2023-12-09T12:07:53.000Z","transactionHash":"0x806dc8d89fc245d5d65ab3ff93b6ee06ef5ce7a47eda739519687b73d3382b45","blockHash":"0x51c027a455ccbf0388175824eff6b6c56fbda36a2e7bb4b69f4d593355e9c4db","codeHash":"0x194b1d7a9d06e6ac0f766532ef58c9ec8004d0f5718a932412755feb9a886f82","chainId":1337,"upgradeable":true}}}}
