+++ hardhat.config.ts
@@ -2,81 +2,18 @@ import "@nomicfoundation/hardhat-toolbox";
 import "@openzeppelin/hardhat-upgrades";
 import "hardhat-contract-sizer";
 // import { ethers } from "hardhat"; //! Cannot be imported here or any file that is imported here because it is generated here
-import { subtask, task, types } from "hardhat/config";
-import { HardhatRuntimeEnvironment, HardhatUserConfig } from "hardhat/types";
-import { HDNodeWallet } from "ethers";
-import { BLOCKCHAIN, GAS_OPT, KEYSTORE } from "configuration";
-import { deploy, upgrade } from "scripts/deploy";
-import { Deploy, SignerInformation, Upgrade } from "models/Tasks";
-import JSON5 from "json5";
-import CustomWallet from "models/Wallet";
-import Environment, { networkNameToId } from "models/Configuration";
-import { delay } from "scripts/utils";
+import { HardhatUserConfig } from "hardhat/types";
+import { BLOCKCHAIN, KEYSTORE } from "configuration";
+import { networkNameToId } from "models/Configuration";
+import { quickTest, createSigner, deploy, upgrade } from "tasks/index";
+
 //* TASKS
-import { hello } from "tasks/example";
-
-hello;
-
-subtask("create-signer", "Creates new signer from given params")
-  // Signer params
-  .addOptionalParam(
-    "relativePath",
-    "Path relative to KEYSTORE.root to store the wallets",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "password",
-    "Password to decrypt the wallet",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "privateKey",
-    "A private key in hexadecimal can be used to sign",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "mnemonicPhrase",
-    "Mnemonic phrase to generate wallet from",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "mnemonicPath",
-    "Mnemonic path to generate wallet from",
-    undefined,
-    types.string,
-  )
-  .setAction(async (args: SignerInformation, hre) => {
-    const { provider } = new Environment(hre);
-    let wallet: CustomWallet | HDNodeWallet | undefined;
-    if (args.privateKey) {
-      wallet = new CustomWallet(args.privateKey, provider);
-    } else if (args.mnemonicPhrase) {
-      wallet = CustomWallet.fromPhrase(
-        args.mnemonicPhrase.toLowerCase() === "default"
-          ? undefined
-          : args.mnemonicPhrase,
-        provider,
-        args.mnemonicPhrase.toLowerCase() === "default"
-          ? undefined
-          : args.mnemonicPath,
-      );
-    } else if (args.relativePath) {
-      wallet = CustomWallet.fromEncryptedJsonSync(
-        args.relativePath,
-        args.password,
-      );
-    } else {
-      throw new Error(
-        "❌  Cannot get a wallet from parameters, needed private key or mnemonic or path",
-      );
-    }
-    return wallet;
-  });
-
+quickTest;
+// WALLET
+createSigner;
+// DEPLOY
+deploy;
+upgrade;
 // task("generate-wallets", "Generates Encryped JSON persistent wallets")
 //   .addPositionalParam(
 //     "type",
@@ -246,206 +183,6 @@ subtask("create-signer", "Creates new signer from given params")
 //   });
 
 // DEPLOYMENTS
-task("deploy", "Deploy smart contracts on '--network'")
-  .addPositionalParam(
-    "contractName",
-    "Name of the contract to deploy (main use: get factory)",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "contractArgs",
-    "(Optional) [undefined] Contract initialize function's arguments if any",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "tag",
-    "(Optional) [undefined] string to include metadata or anything related with a deployment",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "txValue",
-    "Contract creation transaction value if any",
-    undefined,
-    types.int,
-  )
-  // Signer params
-  .addOptionalParam(
-    "relativePath",
-    "Path relative to KEYSTORE.root to store the wallets",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "password",
-    "Password to decrypt the wallet",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "privateKey",
-    "A private key in hexadecimal can be used to sign",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "mnemonicPhrase",
-    "Mnemonic phrase to generate wallet from",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "mnemonicPath",
-    "Mnemonic path to generate wallet from",
-    undefined,
-    types.string,
-  )
-  .setAction(async (args: Deploy, hre) => {
-    console.log(
-      "\x1b[33m ❗️ WARNING: Remember to compile contracts if needed\x1b[0m",
-    );
-    const env = new Environment(hre);
-    const deployer = (await hre.run("create-signer", {
-      relativePath: args.relativePath,
-      password: args.password,
-      privateKey: args.privateKey,
-      mnemonicPhrase: args.mnemonicPhrase,
-      mnemonicPath: args.mnemonicPath,
-    } as SignerInformation)) as CustomWallet | HDNodeWallet;
-    const result = await deploy(
-      args.contractName,
-      deployer,
-      args.contractArgs ? JSON5.parse(args.contractArgs as string) : [],
-      {
-        ...GAS_OPT.max,
-        value: args.txValue,
-      },
-      true,
-      args.tag,
-    );
-    //* Print Result on screen
-    console.info(
-      "\n✅ 🎉 Contract deployed successfully! Contract Information:",
-      `\n  - Contract Name (id within this project): ${result.name}`,
-      `\n  - Contract Address (proxy address if upgradeable deployment): ${result.address}`,
-      result.upgradeable
-        ? `\n  - Logic | Implementation Address (only if upgradeable deployment): ${result.logic}`
-        : "",
-      `\n  - Admin or Deployer: ${deployer.address}`,
-      `\n  - Deploy Timestamp: ${result.timestamp}`,
-      `\n  - Bytecode Hash: ${await result.codeHash()}`,
-      `\n  - Tag: ${result.tag}`,
-    );
-  });
-
-task("upgrade", "Upgrade a deployed smart contracts on '--network'")
-  .addPositionalParam(
-    "contractName",
-    "Name of the contract to upgrade (main use: get factory)",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "address",
-    "(Optional) [undefined] Contract address or proxy address",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "logic",
-    "(Optional) [undefined] Contract logic address or implementation address",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "contractArgs",
-    "(Optional) [undefined] Contract initialize function's arguments if any",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "tag",
-    "(Optional) [undefined] string to include metadata or anything related with a deployment",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "txValue",
-    "Contract creation transaction value if any",
-    undefined,
-    types.int,
-  )
-  // Signer params
-  .addOptionalParam(
-    "relativePath",
-    "Path relative to KEYSTORE.root to store the wallets",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "password",
-    "Password to decrypt the wallet",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "privateKey",
-    "A private key in hexadecimal can be used to sign",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "mnemonicPhrase",
-    "Mnemonic phrase to generate wallet from",
-    undefined,
-    types.string,
-  )
-  .addOptionalParam(
-    "mnemonicPath",
-    "Mnemonic path to generate wallet from",
-    undefined,
-    types.string,
-  )
-  .setAction(async (args: Upgrade, hre) => {
-    console.log(
-      "\x1b[33m ❗️ WARNING: Remember to compile contracts if needed\x1b[0m",
-    );
-    const env = new Environment(hre);
-    const upgrader = (await hre.run("create-signer", {
-      relativePath: args.relativePath,
-      password: args.password,
-      privateKey: args.privateKey,
-      mnemonicPhrase: args.mnemonicPhrase,
-      mnemonicPath: args.mnemonicPath,
-    } as SignerInformation)) as CustomWallet | HDNodeWallet;
-    const result = await upgrade(
-      args.contractName,
-      upgrader,
-      args.contractArgs ? JSON5.parse(args.contractArgs as string) : [],
-      {
-        ...GAS_OPT.max,
-        value: args.txValue,
-      },
-      args.address,
-      args.logic,
-      args.tag,
-    );
-    //* Print Result on screen
-    console.info(
-      "\n✅ 🎉 Contract upgraded successfully! Contract Information:",
-      `\n  - Contract Name (id within this project): ${result.name}`,
-      `\n  - Contract Address (proxy address if upgradeable deployment): ${result.address}`,
-      result.upgradeable
-        ? `\n  - Logic | Implementation Address (only if upgradeable deployment): ${result.logic}`
-        : "",
-      `\n  - Admin or Deployer: ${upgrader.address}`,
-      `\n  - Deploy Timestamp: ${result.timestamp}`,
-      `\n  - Bytecode Hash: ${await result.codeHash()}`,
-      `\n  - Tag: ${result.tag}`,
-    );
-  });
 
 // task(
 //   "call-contract",
@@ -842,41 +579,6 @@ task("upgrade", "Upgrade a deployed smart contracts on '--network'")
 //       ),
 //     );
 //   });
-
-task("quick-test", "Random quick testing function")
-  .addOptionalParam(
-    "args",
-    "Contract initialize function's arguments if any",
-    undefined,
-    types.json,
-  )
-  .setAction(async ({ args }, hre: HardhatRuntimeEnvironment) => {
-    if (args) {
-      // example: npx hardhat quick-test --args '[12, "hello"]'
-      console.log(
-        "RAW Args: ",
-        args,
-        typeof args[0],
-        args[0],
-        typeof args[1],
-        args[1],
-      );
-    }
-    const env = new Environment(hre);
-    console.log(env);
-    console.log("Latest block: ", await hre.ethers.provider.getBlockNumber());
-    console.log(
-      "First accounts: ",
-      await (await hre.ethers.provider.getSigner(0)).getAddress(),
-      await (await hre.ethers.provider.getSigner(1)).getAddress(),
-    );
-    console.log(
-      "First account balance: ",
-      await hre.ethers.provider.getBalance(
-        await (await hre.ethers.provider.getSigner(0)).getAddress(),
-      ),
-    );
-  });
 
 //* Config
 // You need to export an object to set up your config
