+++ package.json
@@ -30,7 +30,6 @@
     "@openzeppelin/hardhat-upgrades": "^2.5.0",
     "@types/chai": "^4.3.11",
     "@types/node": "^20.10.4",
-    "bun-types": "^1.0.15",
     "chai": "^4.3.10",
     "ganache": "^7.9.1",
     "hardhat": "^2.19.2",
@@ -39,7 +38,6 @@
     "mocha-steps": "^1.3.0",
     "prettier": "^3.1.0",
     "prettier-plugin-solidity": "^1.2.0",
-    "tsconfig-paths": "^4.2.0",
-    "yesno": "^0.4.0"
+    "tsconfig-paths": "^4.2.0"
   }
 }
