--- tasks/example.ts
@@ -1,5 +1,0 @@
-import { task } from "hardhat/config";
-
-export const hello = task("hello", "Prints 'Hello, World!'", async () => {
-  console.log("Hello, World!");
-});
