+++ tasks/index.ts
@@ -1,0 +1,3 @@
+export { quickTest } from "./test";
+export { createSigner } from "./wallet";
+export { deploy, upgrade } from "./deploy";
